% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerOralMicrobiome-package.R
\docType{package}
\name{OralCancerMicrobiome-package}
\alias{CancerOralMicrobiome}
\alias{CancerOralMicrobiome-package}
\alias{OralCancerMicrobiome-package}
\alias{OralCancerMicrobiome}
\title{Dataset Accompanying the article, "Oral microbial signatures of head and neck cancer patients with diverse longitudinal oral mucositis severity patterns"}
\description{
These data can be use to reproduce the results in the paper, "Oral microbial
signatures of head and neck cancer patients with diverse longitudinal oral
mucositis severity patterns," which analyzes a data set originally generated in
"Influence of oral microbiome on longitudinal patterns of oral mucositis
severity in patients with squamous cell carcinoma of the head and neck." Code
to reproduce the examples can be found at the OMScoreAnalysis repository on
Github: https://github.com/SarithaKodikara/OMScoreAnalysis. The RDS object
zhang2023-ps.rds is a phyloseq object that combines all other microbiome and
sample information. It is the main export of this package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://support.bioconductor.org/t/OralCancerMicrobiome}
}

}
\author{
Kris Sankaran
}
