% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhylobarData-package.R
\docType{package}
\name{PhylobarData-package}
\alias{PhylobarData}
\alias{PhylobarData-package}
\title{PhylobarData: Datasets for Microbiome and Immunology Research}
\description{
This package provides datasets used in multiple research studies:
}
\details{
\itemize{
\item \code{CancerOralMicrobiome}: Data from "Oral microbial signatures of head and neck
cancer patients with diverse longitudinal oral mucositis severity patterns"
\item \code{CovidImmunology}: COVID-19 immunology dataset from "Su, Y.,
Chen, D., Yuan, D., et al. (2020). Multi-Omics Resolves a Sharp Disease-State
Shift between Mild and Moderate COVID-19. Cell, 183(6), 1479-1495"
}

Code to reproduce examples can be found at the OMScoreAnalysis repository:
https://github.com/SarithaKodikara/OMScoreAnalysis
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://support.bioconductor.org/t/OralCancerMicrobiome}
}

}
\author{
\strong{Maintainer}: Kris Sankaran \email{ksankaran@wisc.edu} (\href{https://orcid.org/0000-0002-9415-1971}{ORCID}) [funder]

Authors:
\itemize{
  \item Megan Kuo
  \item Jiadong Mao
  \item Saritha Kodikara
}

}
\keyword{internal}
