% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{su2019.covid_immunology}
\alias{su2019.covid_immunology}
\title{Su et al. COVID-19 immunology dataset}
\source{
https://doi.org/10.1016/j.cell.2020.10.037
}
\usage{
su2019.covid_immunology()
}
\value{
A SummarizedExperiment object
}
\description{
SummarizedExperiment containing COVID-19 immunology data from the study
"Multi-Omics Resolves a Sharp Disease-State Shift between Mild and Moderate
COVID-19" by Su et al. 2019, Cell Vol 183 Issue 16.
}
\examples{
\dontrun{
  se <- su2019.covid_immunology()
}
}
\seealso{
https://zenodo.org/records/17430294
}
