% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{zhang2023.oral_microbiome}
\alias{zhang2023.oral_microbiome}
\title{Zhang et al. (2023) oral microbiome dataset}
\source{
Zhang et al. (2023) "Influence of oral microbiome on longitudinal patterns"
}
\usage{
zhang2023.oral_microbiome(metadata = FALSE)
}
\value{
A phyloseq object combining microbiome and sample information
}
\description{
phyloseq object containing oral microbiome data from head and neck
cancer patients with longitudinal oral mucositis severity patterns.
}
\details{
These data can be used to reproduce the results in the paper, "Oral microbial
signatures of head and neck cancer patients with diverse longitudinal oral
mucositis severity patterns," which analyzes a data set originally generated in
"Influence of oral microbiome on longitudinal patterns of oral mucositis
severity in patients with squamous cell carcinoma of the head and neck." Code
to reproduce the examples can be found at the OMScoreAnalysis repository on
Github: https://github.com/SarithaKodikara/OMScoreAnalysis.
}
\examples{
\dontrun{
  ps <- zhang2023.oral_microbiome()
}
}
